// awsx/schema-types.ts
export abstract class Image<TData = any> extends (pulumi.ComponentResource)<TData> {
    public imageUri!: string | pulumi.Output<string>;
    constructor(name: string, args: pulumi.Inputs, opts: pulumi.ComponentResourceOptions = {}) {
        super("awsx:ecr:Image", name, opts.urn ? { imageUri: undefined } : { name, args, opts }, opts);
    }
}
export interface ImageArgs {
    readonly args?: pulumi.Input<Record<string, pulumi.Input<string>>>;
    readonly builderVersion?: BuilderVersionInputs;
    readonly cacheFrom?: pulumi.Input<pulumi.Input<string>[]>;
    readonly cacheTo?: pulumi.Input<pulumi.Input<string>[]>;
    readonly context?: pulumi.Input<string>;
    readonly dockerfile?: pulumi.Input<string>;
    readonly imageName?: pulumi.Input<string>;
    readonly imageTag?: pulumi.Input<string>;
    readonly platform?: pulumi.Input<string>;
    readonly registryId?: pulumi.Input<string>;
    readonly repositoryUrl: pulumi.Input<string>;
    readonly target?: pulumi.Input<string>;
}
// ...
export type BuilderVersionInputs = "BuilderV1" | "BuilderBuildKit";
export type BuilderVersionOutputs = "BuilderV1" | "BuilderBuildKit";
export interface DockerBuildInputs {
    readonly args?: pulumi.Input<Record<string, pulumi.Input<string>>>;
    readonly builderVersion?: BuilderVersionInputs;
    readonly cacheFrom?: pulumi.Input<pulumi.Input<string>[]>;
    readonly cacheTo?: pulumi.Input<pulumi.Input<string>[]>;
    readonly context?: pulumi.Input<string>;
    readonly dockerfile?: pulumi.Input<string>;
    readonly imageName?: pulumi.Input<string>;
    readonly imageTag?: pulumi.Input<string>;
    readonly platform?: pulumi.Input<string>;
    readonly target?: pulumi.Input<string>;
}